[enhancement-3481] Filing: 2 Registered advocate names are not showing to add them in the case (#2284)
