+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/dristi/src/Module.js
@@ -68,6 +68,7 @@ import CustomTextInput from "./components/CustomTextInput";
 import CustomEmailTextInput from "./pages/citizen/registration/CustomEmailTextInput";
 import OrSeparator from "./components/OrSeparator";
 import ShowAllTranscriptModal from "./components/ShowAllTranscriptModal";
+import SearchableDropdown from "./components/SearchableDropdown";
 
 export const DRISTIModule = ({ stateCode, userType, tenants }) => {
   const Digit = useMemo(() => window?.Digit || {}, []);
@@ -187,6 +188,7 @@ const componentsToRegister = {
   LogoutIcon,
   TriangleIcon,
   ShowAllTranscriptModal,
+  SearchableDropdown,
 };
 
 const overrideHooks = () => {
